# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

The existing `bucket_internal`  for MPT model implements the processing of the first token. This PR supplements the processing of subsequent tokens. Although throughput has been improved, the changing shape of `key_states` and `value_states` has increased memory usage. 

I am exploring whether dynamic shape can solve this problem. If anyone has any clues, please let me know. Thanks!

```bash
python run_generation.py --model_name_or_path mosaicml/mpt-7b --use_hpu_graphs --use_kv_cache --limit_hpu_graph --batch_size 128  --max_input_tokens 128 --max_new_tokens 128 --trim_logits --attn_softmax_bf16 --warmup 3 --n_iterations 1 --bf16 --bucket_internal --bucket_size 32
```
Test results:
|bs/max_input/max_output|Throughput(including tokenization)(tokens/s)|Memory(GB)|Max_Memory(GB)|
|:---:|:---:|:---:|:---:|
|before|128/128/128|5996|37.04|61.09|
|after|128/128/128|6443|45.07|62.62|
|before|16/128/1024|1322|26.4|36.11|
|after|16/128/1024|1673|30.93|40.12|
|before|32/128/512|2527|28.02|39.83|
|after|32/128/512|2997|33.08|43.46|
|before|64/128/256|4263|31.28|47.52|
|after|64/128/256|4816|36.94|49.72|

## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?