@Jing1Ling 2024-08-28
The issue of increased memory usage was resolved by commenting the code that converts past_key_values ​​from a list to a tuple.

In addition, I set kv_cache_len as an attribute of the attention class when preparing the model input for the first token generation, instead of passing it to the attention forward function.

**Test Result:**
```bash
python run_generation.py --model_name_or_path mosaicml/mpt-7b --use_hpu_graphs --use_kv_cache --limit_hpu_graph --batch_size 128  --max_input_tokens 128 --max_new_tokens 128 --trim_logits --attn_softmax_bf16 --warmup 3 --n_iterations 1 --bf16 --bucket_internal --bucket_size 32
```
| Version | bucket_size |bs/max_input/max_output | Throughput(including tokenization)(tokens/s) | Memory(GB) | Max_Memory(GB)|
| :--: | :--: | :--: | :--: | :--: | :--: |
|before|/|128/128/128|5997|37.03|58.55|
|after|128|128/128/128|6001|29.04|46.6|
|after|32|128/128/128|6441|29.07|46.63|
|before|/|16/128/1024|1322|26.37|35.76|
|after|128|16/128/1024|1638|21.88|31.07|
|after|32|16/128/1024|1671|21.91|31.11|
|before|/|32/128/512|2525|28.02|38.95|
|after|128|32/128/512|2899|23.03|33.43|
|after|32|32/128/512|2998|23.07|33.46|
|before|/|64/128/256|4260|31.15|45.6|
|after|128|64/128/256|4599|24.91|37.68|
|after|32|64/128/256|4811|24.94|37.72|

<!--EndFragment-->
</body>
</html>